+++ cross/auto_parameters/clean_data/missing_values_handler.py
@@ -37,7 +37,7 @@ class MissingValuesParamCalculator:
         if not columns_with_nulls:
             return None
 
-        best_handling_options = {}
+        best_transformation_options = {}
         best_n_neighbors = {}
 
         for column in tqdm(columns_with_nulls, disable=(not verbose)):
@@ -52,12 +52,12 @@ class MissingValuesParamCalculator:
                 column,
                 is_num_column=(column in num_columns),
             )
-            best_handling_options[column] = best_strategy
+            best_transformation_options[column] = best_strategy
 
             if best_strategy == "fill_knn":
                 best_n_neighbors.update(best_params)
 
-        return self._build_result(best_handling_options, best_n_neighbors)
+        return self._build_result(best_transformation_options, best_n_neighbors)
 
     def _get_columns_with_nulls(self, x):
         return x.columns[x.isnull().any()].tolist()
@@ -134,18 +134,18 @@ class MissingValuesParamCalculator:
         strategy,
         n_neighbors=None,
     ):
-        handling_options = {column: strategy}
+        transformation_options = {column: strategy}
         knn_params = {column: n_neighbors} if n_neighbors else None
 
         missing_values_handler = MissingValuesHandler(
-            handling_options=handling_options, n_neighbors=knn_params
+            transformation_options=transformation_options, n_neighbors=knn_params
         )
 
         return evaluate_model(x, y, model, scoring, cv, groups, missing_values_handler)
 
-    def _build_result(self, best_handling_options, best_n_neighbors):
+    def _build_result(self, transformation_options, n_neighbors):
         missing_values_handler = MissingValuesHandler(
-            handling_options=best_handling_options, n_neighbors=best_n_neighbors
+            transformation_options=transformation_options, n_neighbors=n_neighbors
         )
         return {
             "name": missing_values_handler.__class__.__name__,