+++ cross/auto_transform.py
@@ -6,45 +6,9 @@ from typing import Callable, List, Optional, Union
 import numpy as np
 
 import cross.auto_parameters as pc
+from cross.transformations import ColumnSelection
 from cross.transformations.utils.dtypes import numerical_columns
 from cross.utils import get_transformer
-
-
-def date_time() -> str:
-    return datetime.now().strftime("%Y/%m/%d %H:%M:%S")
-
-
-def execute_transformation(
-    calculator,
-    X,
-    y,
-    model,
-    scoring,
-    direction,
-    cv,
-    groups,
-    verbose,
-    transformations,
-    subset=None,
-):
-    if verbose:
-        print(
-            f"\n[{date_time()}] Fitting transformation: {calculator.__class__.__name__}"
-        )
-
-    initial_columns = set(X.columns)
-    X_subset = X.loc[:, subset] if subset else X
-
-    transformation = calculator.calculate_best_params(
-        X_subset, y, model, scoring, direction, cv, groups, verbose
-    )
-
-    if transformation:
-        transformations.append(transformation)
-        transformer = get_transformer(transformation["name"], transformation["params"])
-        X = transformer.fit_transform(X, y)
-
-    return X, list(set(X.columns) - initial_columns)
 
 
 def auto_transform(
@@ -80,8 +44,10 @@ def auto_transform(
         print(f"[{date_time()}] Scoring: {scoring}\n")
 
     X = X.copy()
+    initial_columns = set(X.columns)
     initial_num_columns = numerical_columns(X)
     transformations = []
+    tracked_columns = []
 
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore")
@@ -146,10 +112,125 @@ def auto_transform(
                 groups,
                 verbose,
                 transformations,
+                tracked_columns,
                 subset_columns,
             )
+            final_columns = set(X.columns)
 
             if key:
                 additional_columns[key] = new_columns
 
+    transformations = filter_transformations(
+        transformations, tracked_columns, initial_columns, final_columns
+    )
     return transformations
+
+
+def date_time() -> str:
+    return datetime.now().strftime("%Y/%m/%d %H:%M:%S")
+
+
+def execute_transformation(
+    calculator,
+    X,
+    y,
+    model,
+    scoring,
+    direction,
+    cv,
+    groups,
+    verbose,
+    transformations,
+    tracked_columns,
+    subset=None,
+):
+    if verbose:
+        print(
+            f"\n[{date_time()}] Fitting transformation: {calculator.__class__.__name__}"
+        )
+
+    initial_columns = set(X.columns)
+    X_subset = X.loc[:, subset] if subset else X
+
+    transformation = calculator.calculate_best_params(
+        X_subset, y, model, scoring, direction, cv, groups, verbose
+    )
+
+    if transformation:
+        transformations.append(transformation)
+        transformer = get_transformer(
+            transformation["name"], transformation["params"] | {"track_columns": True}
+        )
+        X = transformer.fit_transform(X, y)
+
+        tracked_columns.append(transformer.tracked_columns)
+
+    return X, list(set(X.columns) - initial_columns)
+
+
+def filter_transformations(
+    transformations, column_dependencies, initial_columns, target_columns
+):
+    filtered_transformations = []
+    required_columns = set(target_columns)
+
+    for index in range(len(transformations) - 1, -1, -1):
+        transformation = transformations[index]
+        transformation_name = transformation["name"]
+        transformation_params = transformation["params"].copy()
+
+        dependency_mapping = column_dependencies[index]
+        additional_required_columns = {
+            source_col
+            for output_col, source_cols in dependency_mapping.items()
+            if output_col in required_columns
+            for source_col in source_cols
+        }
+        required_columns.update(additional_required_columns)
+
+        modified_param_key = None
+
+        if "features" in transformation_params:
+            transformation_params["features"] = [
+                col
+                for col in transformation_params["features"]
+                if col in required_columns
+            ]
+            modified_param_key = "features"
+
+        elif "transformation_options" in transformation_params:
+            transformation_params["transformation_options"] = {
+                key: value
+                for key, value in transformation_params[
+                    "transformation_options"
+                ].items()
+                if key in required_columns
+            }
+            modified_param_key = "transformation_options"
+
+        elif "operations_options" in transformation_params:
+            transformation_params["operations_options"] = [
+                (col1, col2, op)
+                for col1, col2, op in transformation_params["operations_options"]
+                if col1 in required_columns and col2 in required_columns
+            ]
+            modified_param_key = "operations_options"
+
+        if modified_param_key and transformation_params[modified_param_key]:
+            filtered_transformations.append(
+                {
+                    "name": transformation_name,
+                    "params": transformation_params,
+                }
+            )
+
+    # Add column selector to minimize initial columns
+    selected_columns = [col for col in initial_columns if col in required_columns]
+    column_selector = ColumnSelection(selected_columns)
+    selector_transformation = {
+        "name": column_selector.__class__.__name__,
+        "params": column_selector.get_params(),
+    }
+    filtered_transformations.append(selector_transformation)
+
+    return filtered_transformations[::-1]