+++ cross/transformations/numerical_features/normalization.py
@@ -3,9 +3,11 @@ from sklearn.preprocessing import Normalizer
 
 
 class Normalization(BaseEstimator, TransformerMixin):
-    def __init__(self, transformation_options=None):
+    def __init__(self, transformation_options=None, track_columns=False):
         self.transformation_options = transformation_options or {}
+        self.track_columns = track_columns
 
+        self.tracked_columns = {}
         self._transformers = {}
 
     def get_params(self, deep=True):
@@ -33,6 +35,9 @@ class Normalization(BaseEstimator, TransformerMixin):
 
         for column, transformer in self._transformers.items():
             X[column] = transformer.transform(X[[column]])
+
+            if self.track_columns:
+                self.tracked_columns[column] = [column]
 
         return X