+++ cross/transformations/numerical_features/quantile_transformation.py
@@ -3,9 +3,11 @@ from sklearn.preprocessing import QuantileTransformer
 
 
 class QuantileTransformation(BaseEstimator, TransformerMixin):
-    def __init__(self, transformation_options=None):
+    def __init__(self, transformation_options=None, track_columns=False):
         self.transformation_options = transformation_options or {}
+        self.track_columns = track_columns
 
+        self.tracked_columns = {}
         self._transformers = {}
         self._n_quantiles = 1000
 
@@ -41,6 +43,9 @@ class QuantileTransformation(BaseEstimator, TransformerMixin):
 
         for column, transformer in self._transformers.items():
             X[column] = transformer.transform(X[[column]])
+
+            if self.track_columns:
+                self.tracked_columns[column] = [column]
 
         return X