+++ cross/transformations/periodic_features/cyclical_features_transformer.py
@@ -3,12 +3,15 @@ from sklearn.base import BaseEstimator, TransformerMixin
 
 
 class CyclicalFeaturesTransformer(BaseEstimator, TransformerMixin):
-    def __init__(self, columns_periods=None):
-        self.columns_periods = columns_periods or {}
+    def __init__(self, transformation_options=None, track_columns=False):
+        self.transformation_options = transformation_options or {}
+        self.track_columns = track_columns
+
+        self.tracked_columns = {}
 
     def get_params(self, deep=True):
         return {
-            "columns_periods": self.columns_periods,
+            "transformation_options": self.transformation_options,
         }
 
     def set_params(self, **params):
@@ -23,9 +26,13 @@ class CyclicalFeaturesTransformer(BaseEstimator, TransformerMixin):
     def transform(self, X, y=None):
         X = X.copy()
 
-        for column, period in self.columns_periods.items():
+        for column, period in self.transformation_options.items():
             X[f"{column}_sin"] = np.sin(2 * np.pi * X[column] / period)
             X[f"{column}_cos"] = np.cos(2 * np.pi * X[column] / period)
+
+            if self.track_columns:
+                self.tracked_columns[f"{column}_sin"] = [column]
+                self.tracked_columns[f"{column}_cos"] = [column]
 
         return X